+++ index.d.ts
@@ -453,6 +453,8 @@ declare module 'pc-nrfconnect-shared' {
 
     // AppTypes.ts
 
+    export type SourceJson = import('./src/utils/AppTypes').SourceJson;
+    export type AppJson = import('./src/utils/AppTypes').AppJson;
     export type PackageJson = import('./src/utils/AppTypes').PackageJson;
 
     // useHotKey.js